4_Code/scripts/python/requirements.txt
traci == 1.15.0
sumolib == 1.15.0
matplotlib >= 3.5.3
PySimpleGUI == 4.60.4
pillow >= 6.2.0
numpy == 1.21.6
pandas == 1.3.5
joblib == 1.2.0
PyYAML == 6.0
typing-extensions == 4.4.0
folium == 0.14.0

# Development Linter and Formatter
flake8 >= 5.0.4
flake8-quotes >= 3.3.2
flake8-simplify >= 0.19.3
pep8-naming >= 0.13.3
autopep8 >= 2.0.1
